Move generic_template::render to utils::render
