+++ packages/sycamore/src/generic_node/dom_node.rs
@@ -9,7 +9,7 @@ use wasm_bindgen::prelude::*;
 use wasm_bindgen::{intern, JsCast};
 use web_sys::{Comment, Element, Node, Text};
 
-use crate::generic_node::render::insert;
+use crate::utils::render::insert;
 use crate::generic_node::{EventHandler, GenericNode};
 use crate::rx::{create_root, on_cleanup, ReactiveScope};
 use crate::template::Template;
